+++ packages/kit/src/runtime/server/page/types.d.ts
@@ -1,0 +1,8 @@
+import { LoadOutput, SSRNode } from 'types.internal';
+
+export type Loaded = {
+	node: SSRNode;
+	loaded: LoadOutput;
+	fetched: Array<{ url: string; json: string }>;
+	uses_credentials: boolean;
+};
